Move /common/ folder to /core/ and coremodule
fix up all the reference to activity services and activity classes.
Move into core module.
