+++ src/app/core/activity-data.ts
