+++ src/app/core/activity.ts
